backend/requirements.txt
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.1.8
cloudinary==1.41.0
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.0.0
six==1.17.0
SQLAlchemy==2.0.23
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
